refactor(constants): ActionType定義ファイルをxxxActionTypes.jsに変更
