+++ src/constants/timerActionTypes.js
